code/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "di4d-portal"
version = "0.1.0"
description = "DI4D Portal (Django)"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "Django==5.2.9",
    "Jinja2>=3.1.3",
    "whitenoise>=6.8.2",
    "python-dotenv>=1.0.1",
    "Pillow>=12.0.0",
    "django-oidc-provider>=0.9.0"
    ,"mozilla-django-oidc>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "django-tailwind>=4.4.2",
    "django-livereload-server>=0.5.1"
]
prod = [
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.9"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"" = "DI4D_Portal"

[tool.setuptools.packages.find]
where = ["DI4D_Portal"]
